// modules/PDFView/models/GeneralInfo.ts
export interface GeneralInfo {
  profilePicture: string;
  firstName: string;
  lastName: string;
  dob: string;
  aboutMe: string;
  position: string;
  email: string;
  phone: string;
  address: string;
  city: string;
  zip: string;
  country: string;
  website: string;
  LinkedIn: string;
  GitHub: string;
  Facebook: string;
  Twitter: string;
}